Add nesting layer for codegen tests and add test for input_object derive
