+++ juniper_tests/Cargo.toml
@@ -5,6 +5,7 @@ version = "0.1.0"
 [dependencies]
 juniper = { path = "../juniper" }
 juniper_codegen = { path = "../juniper_codegen" }
+serde_json = { version = "^1.0.2" }
 
 [[test]]
 name = "integration_tests"
