+++ juniper_tests/src/lib.rs
@@ -1,4 +1,5 @@
 #[macro_use] extern crate juniper;
 #[macro_use] extern crate juniper_codegen;
+#[macro_use] extern crate serde_json;
 
 mod codegen;
